+++ src/constants/routes.ts
@@ -1,0 +1,8 @@
+export const ROUTES = {
+  HOME: '/',
+  COURSES: '/courses',
+  STUDY_GROUP: '/study-group',
+  RECRUITMENT: '/recruitment',
+  RECRUITMENT_MANAGE: '/recruitment/manage',
+  LOGIN: '/login',
+} as const
